chore: move resolver under decorator
